// src/app/types/rokuEnum.ts
export enum ModuleStatus {
  DRAFT = 'draft',
  LIVE = 'live',
  READY = 'ready',
  PUBLISH_PROGRESS = 'publish in progress',
  COMPLETE = 'complete',
  ENDED = 'ended',
  ARCHIVED = 'archived',
}

export enum ModuleValueStatus {
  DFT = 'default',
  SAVED = 'saved',
  PUBLISHED = 'published',
  INCOMPLETE = 'incomplete',
  ARCHIVED = 'archived',
  ENDED = 'ended',
}